api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0         
asyncpg==0.30.0         
redis==5.0.0            
bcrypt==4.1.2           
python-dotenv==1.0.0    
httpx==0.27.0           